tests/e2e/requirements.txt
# E2E test dependencies
# These are in addition to the main requirements.txt

# LiveKit SDK for testing
livekit>=0.9.0

# HTTP client for API testing
httpx>=0.25.0

# Additional testing utilities
pytest-timeout>=2.1.0  # For test timeouts
